% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_functions.R
\name{analyse_limit_orders}
\alias{analyse_limit_orders}
\title{analyse_limit_orders}
\usage{
analyse_limit_orders(order_book)
}
\arguments{
\item{order_book}{A populated order book to analyse}
}
\value{

}
\description{
analyse_limit_orders
}
