% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_functions.R
\name{calculate_measures}
\alias{calculate_measures}
\title{calculate_measures}
\usage{
calculate_measures(order_book, fill_times, num_orders)
}
\arguments{
\item{order_book}{The order book to analyse.}

\item{fill_times}{The fill time of the orders (optional).}

\item{num_orders}{The number of orders to analyse (optional).}
}
\value{

}
\description{
calculate_measures
}
