% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_functions.R
\name{create_order_book}
\alias{create_order_book}
\title{create_order_book}
\usage{
create_order_book(
  start_price,
  price_limit,
  price_band,
  tick_size,
  market_sell,
  market_buy,
  bid_order_ratio,
  ask_order_ratio,
  cancel_bid,
  cancel_ask,
  n_events
)
}
\arguments{
\item{n_events}{}
}
\value{

}
\description{
create_order_book
}
