% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_functions.R
\name{fill_order}
\alias{fill_order}
\title{fill_order}
\usage{
fill_order(
  market_order,
  order_book,
  record,
  current_price,
  tick_size,
  next_order_position,
  mm_price
)
}
\arguments{
\item{market_order}{The order to be filled.}

\item{order_book}{The order book}

\item{record}{The trading record (which is tracking the traded prices)}

\item{current_price}{The current trading price}

\item{tick_size}{The minimum tick size allowed}

\item{next_order_position}{The position of the next order in the book}

\item{mm_price}{THe market makers price}
}
\value{

}
\description{
Fill orders once available market depth is available.
}
