% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functons.R
\name{plot_price_series}
\alias{plot_price_series}
\title{plot_series}
\usage{
plot_price_series(price_series)
}
\arguments{
\item{price_series}{The price series to be visualised.}
}
\value{

}
\description{
plot_series
}
