% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functons.R
\name{plot_survival_curves}
\alias{plot_survival_curves}
\title{plot_survival_curves}
\usage{
plot_survival_curves(results_matrix)
}
\arguments{
\item{results_matrix}{}
}
\value{

}
\description{
plot_survival_curves
}
