% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_functions.R
\name{calculate_measures}
\alias{calculate_measures}
\title{calculate_measures}
\usage{
calculate_measures(order.book, fill.times, num.orders)
}
\arguments{
\item{num.orders}{}
}
\value{

}
\description{
calculate_measures
}
