% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_functions.R
\name{create_orders}
\alias{create_orders}
\title{create_orders}
\usage{
create_orders(
  time_stamp,
  order_number,
  market_sell,
  market_buy,
  limit_ask,
  limit_bid,
  cancel_bid,
  cancel_ask
)
}
\arguments{
\item{cancel_ask}{}
}
\value{

}
\description{
create_orders
}
