% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_functions.R
\name{create_orders}
\alias{create_orders}
\title{create_orders}
\usage{
create_orders(
  time.stamp,
  order.number,
  market.sell,
  market.buy,
  limit.ask,
  limit.bid,
  cancel.bid,
  cancel.ask
)
}
\arguments{
\item{cancel.ask}{}
}
\value{

}
\description{
create_orders
}
