% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_functions.R
\name{fill_order}
\alias{fill_order}
\title{fill_order}
\usage{
fill_order(
  market_order,
  order_book,
  record,
  current_price,
  tick_size,
  next_order_position,
  mm_price
)
}
\arguments{
\item{mm_price}{}
}
\value{

}
\description{
fill_order
}
