% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_functions.R
\name{fill_order}
\alias{fill_order}
\title{fill_order}
\usage{
fill_order(
  market.order,
  order.book,
  record,
  current.price,
  tick.size,
  next.order.position,
  cf.price
)
}
\arguments{
\item{cf.price}{}
}
\value{

}
\description{
fill_order
}
